requirements.txt
flask==3.0.2
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
tensorflow==2.15.0
keras==2.15.0
sklearn==1.4.1
flask_wtf==1.2.1
pillow==10.2.0
joblib==1.3.2